Initial commit
At present, it loads a react bundle on / and has an auth system and can
do basic things like register, forgot password and change account details. It has some other pages as well, but they're from the old project.

This is a project stripped from another. I have removed most of the
references to the old project but it's far from perfect.